requirements.txt
numpy==1.15.4
opencv-contrib-python==3.4.5.20
Pillow==5.4.0
pytesseract==0.2.6
